修正get score 錯誤，以及增加 main 的進入點

+++ src/main/java/model/GeneticModel.java
@@ -183,6 +183,6 @@ public class GeneticModel implements AlgorithmModel {
         }
 
         int[] bestObj = current[best];
-        return (int) evaluator.evaluate(bestObj);
+        return evaluator.evaluate(bestObj);
     }
 }
